Update patch set 1
Patch Set 1:

Okay, so as a rule of thumb SELinux works by labeling files and allowing only specific things access to the files it want.

So allowing everything access to unlabeled files totally goes against that principle.

What you need to do is label them (using file_contexts), and then allow that specific domain access to only the files it needs.

Look at some other device for an example.
Try to search for the file getting the denial in file_contexts, and seeing how it's labeled and given permissions, and try to follow that.


Looking at the policies though, a lot of it seems due to you not having a good /data partition.
Have you tried to run restorecon -R (for recursive) on it?
Also try doing a factory reset, and see how the files are labeled.
You could also check another device you have (that has this work done) to see how a "global" not device specific file is labeled, and make sure this device has it fine.

Ping me (or anyone else who's around and familiar with SELinux) on IRC if you need help.

Patch-set: 1