cancro: healtd: the worst thing i could do
Since I switch to healtd many humans and softwares noted that device was not charging fast
as it was excepted (in fact hevc is enabled). After years of research scienciests realized that
it was always bringing in less than 220mah. The problem disappeared if the device booted while it
was connected to the charger.
This issue made users so sad that many of them could not sleep at night and they were not able to be
productive at work, so they got fired... Hopefully researchers have found out a magic version of
xiaomi's chargeonlymode bin that works with marshmallow and no longer giver charging problems.

TL;DR: switch back to xiaomi's chargeonlymode

Change-Id: I165f9e090db30de101abe758c49041cd65c29299